Make back/forward update scan
Changed to route params instead of query param at same time
